// src/app/pages/vendors/dto/model/m-add-vendor.ts
export interface MAddVendor {
  vendorName: string;
  shippingAddress: string;
  contactName: string;
  bankInfo: string;
  productTypes: string;
  city: string;
  zip: string;
  country: string;
  telephone: string;
  email: string;
  wechatId: string;
  bankAccountNumber: string;
  alipay: string;
  brand: string;
}